create class id unit test
